Person testes

+++ src/model/WorkingPerson.cpp
@@ -20,7 +20,7 @@ namespace eatms
 
         const std::string WorkingPerson::getId() const{
             auto tempString = idPrefix_;
-            return tempString + std::to_string(idNum_);
+            return tempString + std::to_string(static_cast<long long int>(idNum_));
         }
 
         void WorkingPerson::updateId(const std::string &idString){
